// scripts/helpers/interfaces.ts
interface ClientStruct {
  RPC: string;
  seed: string;
  prefix: string;
}

interface IbcStruct {
  dstPrefix: string;
  sourceChannel: string;
  sourcePort: string;
  amount: number;
}

interface SwapStruct {
  from: AssetSymbol;
  to: AssetSymbol;
  amount: number;
}

interface DelegationStruct {
  targetAddr: string;
  tokenAmount: number;
  tokenDenom: string;
  validatorAddr: string;
}

interface Asset {
  asset_denom: string;
  wallet_address: string;
  wallet_balance: string;
  weight: string;
  amount_to_send_until_next_epoch: string;
}

interface User {
  asset_list: Asset[];
  is_controlled_rebalancing: boolean;
  day_counter: string;
  deposited_on_current_period: string;
  deposited_on_next_period: string;
}

interface PoolExtracted {
  id: string;
  denom: string;
  price: string;
  symbol: string;
  channel_id: string;
  port_id: string;
}

interface UserExtracted {
  osmo_address: string;
  asset_list: AssetExtracted[];
}

interface AssetExtracted {
  asset_denom: string;
  wallet_address: string;
  wallet_balance: string;
}

interface QueryPoolsAndUsersResponse {
  users: UserExtracted[];
  pools: PoolExtracted[];
}
// ...
type AssetSymbol =
  | "ATOM"
  | "OSMO"
  | "ION"
  | "AKT"
  | "DVPN"
  | "IRIS"
  | "CRO"
  | "XPRT"
  | "REGEN"
  | "NGM"
  | "EEUR"
  | "JUNO"
  | "LIKE"
  | "USTC"
  | "BCNA"
  | "BTSG"
  | "XKI"
  | "SCRT"
  | "MED"
  | "BOOT"
  | "CMDX"
  | "CHEQ"
  | "STARS"
  | "HUAHUA"
  | "LUM"
  | "DSM"
  | "GRAV"
  | "SOMM"
  | "ROWAN"
  | "NETA"
  | "UMEE"
  | "DEC"
  | "PSTAKE"
  | "DAI"
  | "USDC"
  | "MNTL"
  | "WETH"
  | "WBTC"
  | "EVMOS"
  | "TGD"
  | "DOT"
  | "ODIN"
  | "GLTO"
  | "GEO"
  | "BLD"
  | "CUDOS";